+++ src/scenes/Home/index.jsx
@@ -1,21 +1,19 @@
 import React from 'react';
 
-import VideoHeader from '../../components/Video Header';
+import HomeMasthead from './components/Home Masthead';
 import ProjectsSection from './components/Projects Section';
 import BlogSection from './components/Blog Section';
-import MissionSection from './components/Mission Section';
-import DonateSection from '../Donate Section';
+import DonateSection from './components/Donate Section';
 import AboutSection from './components/About Section';
 
 class Home extends React.Component {
   render() {
     return (
       <React.Fragment>
-        <VideoHeader />
-        <MissionSection />
+        <HomeMasthead />
         <ProjectsSection />
-        <BlogSection />
         <AboutSection />
+        <BlogSection />
         <DonateSection />
       </React.Fragment>
     );