+++ src/scenes/Projects/index.jsx
@@ -1,8 +1,12 @@
 import React from 'react';
 import Proptypes from 'prop-types';
 import { connect } from 'react-redux';
-import { loadProjects } from '../../store/projects/projects';
+
 import { loadHeaderImage } from '../../store/base/base';
+import {
+  loadCurrentProjects,
+  loadFutureProjects
+} from '../../store/projects/projects';
 
 import PageHeader from '../../components/Page Header';
 import { Loader } from '../../components/Loader';
@@ -14,8 +18,11 @@ import './styles.scss';
 // Maps Redux dispatch actions to props
 const mapDispatchToProps = dispatch => {
   return {
-    loadProjects: () => {
-      dispatch(loadProjects());
+    loadCurrentProjects: () => {
+      dispatch(loadCurrentProjects());
+    },
+    loadFutureProjects: () => {
+      dispatch(loadFutureProjects());
     },
     loadHeaderImage: () => {
       dispatch(loadHeaderImage('2iuUpTaObaSCw0kcya8Ci'));
@@ -36,53 +43,29 @@ class Projects extends React.Component {
   }
 
   componentDidMount() {
-    const { loadHeaderImage, loadProjects } = this.props;
+    const {
+      loadHeaderImage,
+      loadCurrentProjects,
+      loadFutureProjects
+    } = this.props;
 
     // Dispatches headerImage()
     loadHeaderImage();
 
-    // Dispatches loadProjects()
-    loadProjects();
-  }
+    // Dispatches loadCurrentProjects()
+    loadCurrentProjects();
 
-  sortProjects(projects) {
-    var currentProjects = new Array(),
-      futureProjects = new Array();
-
-    for (var project in projects) {
-      // Assign project object
-      let projectObj = projects[project];
-
-      // Sort current and future projects from currentProject boolean
-      if (projectObj.fields.currentProject) {
-        currentProjects.push(projectObj);
-      } else {
-        futureProjects.push(projectObj);
-      }
-    }
-
-    // Sort current projects by <date></date>
-    currentProjects.sort(function(a, b) {
-      return new Date(b.fields.date) - new Date(a.fields.date);
-    });
-
-    // Returns array of the storted projects
-    return [currentProjects, futureProjects];
+    // Dispatches loadFutureProjects()
+    loadFutureProjects();
   }
 
   render() {
-    const { loading, projects } = this.props.projects;
-    var sortedProjects;
-
-    // When done loading projects
-    if (!loading) {
-      sortedProjects = this.sortProjects(projects);
-    }
+    const { loading, currentProjects, futureProjects } = this.props.projects;
 
     const { headerImage } = this.props.base;
 
     return (
-      <div className=" bg-gray">
+      <div className="bg-gray">
         <PageHeader headerImage={headerImage}>
           <h1>Projects</h1>
         </PageHeader>
@@ -96,7 +79,7 @@ class Projects extends React.Component {
                   <h2 className="projects-section-title">Current Projects</h2>
                 </div>
                 {// Current projects mapping
-                sortedProjects[0].map(({ fields, sys }, i) => (
+                currentProjects.map(({ fields, sys }, i) => (
                   <div className="col-12 col-lg-6 mb-3 mb-lg-0" key={i}>
                     <CurrentProjectItem {...fields} {...sys} />
                   </div>
@@ -107,7 +90,7 @@ class Projects extends React.Component {
                   <h3 className="projects-section-title">Future Projects</h3>
                 </div>
                 {// Future projects mapping
-                sortedProjects[1].map(({ fields, sys }, i) => (
+                futureProjects.map(({ fields, sys }, i) => (
                   <div className="col-12 col-md-6" key={i}>
                     <FutureProjectItem {...fields} {...sys} />
                   </div>