+++ src/store/base/actions.js
@@ -2,11 +2,19 @@
 
 // Import Redux base types
 import {
+    LOAD_ABOUT_ITEMS_SUCCESS,
     LOAD_PLACEHOLDER_SUCCESS,
     LOAD_HEADER_IMAGE_SUCCESS,
     LOAD_SOCIALS_SUCCESS,
-    LOAD_ABOUT_ITEMS_SUCCESS
 } from './types';
+
+// Action that returns successfully retreived socials
+export function loadAboutItemsSuccess(aboutItems) {
+    return {
+        type: LOAD_ABOUT_ITEMS_SUCCESS,
+        aboutItems
+    }
+}
 
 // Action that returns the successfully loaded header image
 export function loadPlaceholderImageSuccess(placeholderImage) {
@@ -30,12 +38,4 @@ export function loadSocialsSuccess(socials) {
         type: LOAD_SOCIALS_SUCCESS,
         socials
     };
-}
-
-// Action that returns successfully retreived socials
-export function loadAboutItemsSuccess(aboutItems) {
-    return {
-        type: LOAD_ABOUT_ITEMS_SUCCESS,
-        aboutItems
-    }
 }