+++ src/store/initialState.js
@@ -5,9 +5,15 @@ export default {
         headerImage: '',
         placeholderImage: ''
     },
+    home: {
+        aboutBlurb: '',
+        donateBlurb: '',
+        loading: false
+    },
     projects: {
         loading: false,
-        projects: [],
+        currentProjects: [],
+        futureProjects: [],
         projectPost: []
     },
     blog: {