+++ src/store/projects/types.js
@@ -3,8 +3,11 @@
 // Loading projects type
 export const PROJECTS_LOADING = 'PROJECTS_LOADING';
 
-// Successfully loaded projects type
-export const LOAD_PROJECTS_SUCCESS = 'LOAD_PROJECTS_SUCCESS';
+// Successfully loaded current projects type
+export const LOAD_CURRENT_PROJECTS_SUCCESS = 'LOAD_CURRENT_PROJECTS_SUCCESS';
+
+// Successfully loaded future projects type
+export const LOAD_FUTURE_PROJECTS_SUCCESS = 'LOAD_FUTURE_PROJECTS_SUCCESS';
 
 // Successfully loaded projects post type
 export const LOAD_PROJECT_POST_SUCCESS = 'LOAD_PROJECT_POST_SUCCESS';