+++ src/store/rootReducer.js
@@ -2,6 +2,7 @@ import {
     combineReducers
 } from 'redux'
 import base from './base/reducer'
+import home from './home/reducer'
 import projects from './projects/reducer'
 import blog from './blog/reducer'
 import team from './team/reducer'
@@ -9,6 +10,7 @@ import team from './team/reducer'
 // Redux root reducers
 export const rootReducer = combineReducers({
     base,
+    home,
     projects,
     blog,
     team